+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # changelog
 
+## 1.0.16 (23/03/2026)
+
+### Added
+
+- Button completato con varianti, stati e token CSS
+- Demo Vite/React con menu laterale stile Storybook
+
 ## 1.0 (08/06/2024)
 
 ### Added